--- internal/bft/init.go
@@ -1,6 +1,0 @@
-// Copyright IBM Corp. All Rights Reserved.
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-
-package bft
