+++ pkg/types/types.go
@@ -1,0 +1,52 @@
+// Copyright IBM Corp. All Rights Reserved.
+//
+// SPDX-License-Identifier: Apache-2.0
+//
+
+package bft
+
+import (
+	"crypto/sha256"
+	"encoding/asn1"
+	"encoding/hex"
+	"fmt"
+)
+
+type Proposal struct {
+	Payload              []byte
+	Header               []byte
+	Metadata             []byte
+	VerificationSequence uint64
+}
+
+type Signature struct {
+	Id    uint64
+	Value []byte
+}
+
+type RequestInfo struct {
+	ID       string
+	ClientID string
+}
+
+func (p Proposal) Digest() string {
+	rawBytes, err := asn1.Marshal(Proposal{
+		VerificationSequence: p.VerificationSequence,
+		Metadata:             p.Metadata,
+		Payload:              p.Payload,
+		Header:               p.Header,
+	})
+
+	if err != nil {
+		panic(fmt.Sprintf("failed marshaling proposal: %v", err))
+	}
+
+	return computeDigest(rawBytes)
+}
+
+func computeDigest(rawBytes []byte) string {
+	h := sha256.New()
+	h.Write(rawBytes)
+	digest := h.Sum(nil)
+	return hex.EncodeToString(digest)
+}
